streamlabs-web/requirements.txt
appdirs==1.4.4
asgiref==3.3.1
distlib==0.3.1
Django==3.1.3
django-admin-rangefilter==0.6.3
django-filter==2.4.0
djangorestframework==3.12.2
filelock==3.0.12
importlib-metadata==3.1.1
Markdown==3.3.3
Pillow==8.0.1
psycopg2==2.8.6
pytz==2020.4
reportlab==3.5.55
six==1.15.0
sqlparse==0.4.1
virtualenv==20.2.1
zipp==3.4.0
